# swagger.yml
openapi: '3.0.2'
info:
  title: WAPL Math SAP API
  version: '0.1.0'
  description: |
    지식 수준, 진단고사, 시험 대비 추천 관련 API
servers:
  - url: http://192.168.153.154:8080/
    description: "example server address"
...
paths:
  /diagnosis/scope:
    get:
      tags: [Diagnosis]
      operationId: DiagnosisScopeGet
      summary: 진단고사 범위 중단원 리스트 및 추천 중단원 제공
      parameters:
        - $ref: "#/components/parameters/UserToken"
      responses:
        200:
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosisScopeResult'
        500:
          description: 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  resultMessage:
                    type: string
                    example: "Error: status code 500" 
...
  /userschedulescope:
    get:
      tags:
        - User
      summary: 학생의 현재 추천 단원 리스트 리턴
      description: 학생의 현재 추천 단원 리스트 리턴
      parameters:
        - $ref: "#/components/parameters/UserToken"
      responses:
        "200":
          description: "성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserScheduleScopeGetResult"
        "500":
          description: "실패"
          $ref: "#/components/responses/500-Generic"
    put:
      tags:
        - User
      summary: 학생 추천 단원 범위 입력
      description: 학생 추천 단원 범위 입력
      parameters:
        - $ref: "#/components/parameters/UserToken"
      requestBody:
        $ref:  "#/components/requestBodies/UserScheduleScope"
      responses:
        "200":
          description: "성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserScheduleScopePutResult"
        "500":
          description: "실패"
          $ref: "#/components/responses/500-Generic"
  /user:
    delete:
      tags:
        - 'User'
      summary: "학생 초기화"
      description: 
        학생의 모든 기록 초기화(삭제)
      parameters:
        - $ref: "#/components/parameters/UserToken"
      responses:
        "200":
          description: "성공 시"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteUserResult"
        "500":
          description: "오류 발생 시"
          $ref: "#/components/responses/500-Generic" 
...
components:
  schemas:
...
    DiagnosisScopeResult:
      properties:
        resultMessage:
          type: string
          example: Successfully returned.
        sectionIdList:
          type: array
          description: 진단 범위 내의 모든 중단원 리스트
          example:
            - "중등-중3-1학-03-03"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
        recommendSectionIdList:
          type: array
          description: 각 내용 영역 별 추천 중단원 5개
          example:
            - "중등-중3-1학-03-03"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
...
    UserScheduleScope:
      type: object
      description: 학생 추천 단원 범위. toDefault=true인 경우, sectionIdList null. toDefault=false인 경우, sectionIdList 필수.
      required:
        - toDefault
        - sectionIdList
      properties:
        toDefault:
          type: boolean
          description: 디폴트 범위로 설정 여부. 폴트 범위 = 현재 주차에 해당하는 단원 ~ 이번 학기 끝 단원.
          example: False
        sectionIdList:
          type: array
          nullable: true
          description: 실력학습 범위 중단원 id 리스트
          items:
            type: string
          example: 
            - "중등-중3-1학-03-03"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
...
    UserScheduleScopePutResult:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Successfully update user schedule scope.
          description: "성공 여부 메시지"
    UserScheduleScopeGetResult:
      type: object
      required:
        - subSectionIdList 
      properties:
        sectionIdList:
          type: array
          description: 추천 단원 중단원 id 리스트. (커리큘럼 순서 sorting)
          items:
            type: string
          example: 
            - "중등-중3-1학-04-01"
            - "중등-중3-1학-04-01"
    DeleteUserResult:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Successfully delete user info.
          description: "성공 여부 메시지"
...
  responses:
    500-Generic:
      description: "500 에러에 대한 generic response"
      content:
        application/json:
          schema:
            type: object
            properties:
              errorCode:
                type: string
                description: "Error code"
                example: "ERR-1001"
              message:
                type: string
                description: "Error message"
                example : "Invalid token ID"
  parameters:
    UserToken:
      in: "header"
      name: "token"
      required: true
      description: "유저의 Access token"
      schema:
        type: "string"
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySUQiOiJta2thbmcifQ.ZjD-h4zkD2t8GsqA9or7F8NAbKVGTGEnBEeNvpI8Z5o"
...
  requestBodies:
...
    UserScheduleScope:
      description: "학생 추천 단원 범위"
      required: true
      content:
        application/json:
          schema:
            $ref:  "#/components/schemas/UserScheduleScope"